let Flux sign its own tags as well
We will get an error when flux stores its state in the release branch as
a tag, flux-sync, unless the tag is also signed with a verified key.

The releases/ and manifests/ branch were in "gitops-jenkins" branch, not
in the repo root. Fix git.path to point to the right location.

Load a signing key in (the signing key will not be stored in the repo,
it is in fluxcd-signing which shows up in the next commit as gitignore)

Enable the first-parent signature verification mode which was not yet
present as a configuration option in the FluxCD chart repo's values.yaml
(it is a bug, and the FluxCD maintainers have already been made aware!)